Remove SparkSQLEnv
