--- sql/service/src/main/scala/org/apache/spark/sql/service/SparkSQLEnv.scala
@@ -1,84 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.service
-
-import java.util.Locale
-
-import org.apache.spark.{SparkConf, SparkContext}
-import org.apache.spark.internal.Logging
-import org.apache.spark.sql.{SparkSession, SQLContext}
-import org.apache.spark.sql.internal.StaticSQLConf.CATALOG_IMPLEMENTATION
-import org.apache.spark.util.Utils
-
-/** A singleton object for the master program. The slaves should not access this. */
-private[service] object SparkSQLEnv extends Logging {
-  logDebug("Initializing SparkSQLEnv")
-
-  var sqlContext: SQLContext = _
-  var sparkContext: SparkContext = _
-
-  def init(): Unit = {
-    if (sqlContext == null) {
-      val sparkConf = new SparkConf(loadDefaults = true)
-      // If user doesn't specify the appName, we want to get [SparkSQL::localHostName] instead of
-      // the default appName [SparkSQLCLIDriver] in cli or beeline.
-      val maybeAppName = sparkConf
-        .getOption("spark.app.name")
-        .filterNot(_ == classOf[SparkThriftServer].getName)
-
-      sparkConf
-        .setAppName(maybeAppName.getOrElse(s"SparkSQL::${Utils.localHostName()}"))
-
-      val builder = SparkSession.builder().config(sparkConf)
-      val sparkSession = if (sparkConf.get(CATALOG_IMPLEMENTATION.key, "hive")
-        .toLowerCase(Locale.ROOT) == "hive") {
-        if (SparkSession.hiveClassesArePresent) {
-          builder.enableHiveSupport().getOrCreate()
-        } else {
-          builder.config(CATALOG_IMPLEMENTATION.key, "in-memory")
-          builder.getOrCreate()
-        }
-      } else {
-        builder.getOrCreate()
-      }
-      sparkContext = sparkSession.sparkContext
-      sqlContext = sparkSession.sqlContext
-
-      // SPARK-29604: force initialization of the session state with the Spark class loader,
-      // instead of having it happen during the initialization of the Hive client (which may use a
-      // different class loader).
-      sparkSession.sessionState
-    }
-  }
-
-  def setSQLContext(sqlContext: SQLContext): Unit = {
-    this.sqlContext = sqlContext
-    this.sparkContext = sqlContext.sparkContext
-  }
-
-  /** Cleans up and shuts down the Spark SQL environments. */
-  def stop(): Unit = {
-    logDebug("Shutting down Spark SQL Environment")
-    // Stop the SparkContext
-    if (SparkSQLEnv.sparkContext != null) {
-      sparkContext.stop()
-      sparkContext = null
-      sqlContext = null
-    }
-  }
-}
